Update types and move useAlerts code to useWindows composable
